requirements.txt
streamlit
supabase
requests
pandas
plotly
python-dotenv
APScheduler
